+++ packages/carp_connectivity_package/example/lib/example.dart
@@ -1,5 +1,6 @@
-import 'package:carp_connectivity_package/connectivity.dart';
+import 'package:carp_core/carp_core.dart';
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
+import 'package:carp_connectivity_package/connectivity.dart';
 
 /// This is a very simple example of how this sampling package is used with
 /// CARP Mobile Sensing (CAMS).
@@ -9,14 +10,11 @@ void main() async {
   // register this sampling package before using its measures
   SamplingPackageRegistry().register(ConnectivitySamplingPackage());
 
-  // Create a study protocol using a local file to store data
-  CAMSStudyProtocol protocol = CAMSStudyProtocol()
-    ..name = 'Track patient movement'
-    ..owner = ProtocolOwner(
-      id: 'AB',
-      name: 'Alex Boyon',
-      email: 'alex@uni.dk',
-    );
+  // Create a study protocol
+  StudyProtocol protocol = StudyProtocol(
+    ownerId: 'owner@dtu.dk',
+    name: 'Context Sensing Example',
+  );
 
   // define which devices are used for data collection
   // in this case, its only this smartphone