+++ packages/carp_connectivity_package/lib/connectivity.g.dart
@@ -8,7 +8,7 @@ part of connectivity;
 
 ConnectivityDatum _$ConnectivityDatumFromJson(Map<String, dynamic> json) {
   return ConnectivityDatum()
-    ..id = json['id'] as String
+    ..id = json['id'] as String?
     ..timestamp = json['timestamp'] == null
         ? null
         : DateTime.parse(json['timestamp'] as String)
@@ -26,21 +26,19 @@ Map<String, dynamic> _$ConnectivityDatumToJson(ConnectivityDatum instance) {
 
   writeNotNull('id', instance.id);
   writeNotNull('timestamp', instance.timestamp?.toIso8601String());
-  writeNotNull('connectivity_status', instance.connectivityStatus);
+  val['connectivity_status'] = instance.connectivityStatus;
   return val;
 }
 
 BluetoothDatum _$BluetoothDatumFromJson(Map<String, dynamic> json) {
   return BluetoothDatum()
-    ..id = json['id'] as String
+    ..id = json['id'] as String?
     ..timestamp = json['timestamp'] == null
         ? null
         : DateTime.parse(json['timestamp'] as String)
-    ..scanResult = (json['scan_result'] as List)
-        ?.map((e) => e == null
-            ? null
-            : BluetoothDevice.fromJson(e as Map<String, dynamic>))
-        ?.toList();
+    ..scanResult = (json['scan_result'] as List<dynamic>)
+        .map((e) => BluetoothDevice.fromJson(e as Map<String, dynamic>))
+        .toList();
 }
 
 Map<String, dynamic> _$BluetoothDatumToJson(BluetoothDatum instance) {
@@ -54,23 +52,30 @@ Map<String, dynamic> _$BluetoothDatumToJson(BluetoothDatum instance) {
 
   writeNotNull('id', instance.id);
   writeNotNull('timestamp', instance.timestamp?.toIso8601String());
-  writeNotNull('scan_result', instance.scanResult);
+  val['scan_result'] = instance.scanResult;
   return val;
 }
 
 BluetoothDevice _$BluetoothDeviceFromJson(Map<String, dynamic> json) {
-  return BluetoothDevice()
-    ..advertisementName = json['advertisement_name'] as String
-    ..bluetoothDeviceId = json['bluetooth_device_id'] as String
-    ..bluetoothDeviceName = json['bluetooth_device_name'] as String
-    ..bluetoothDeviceType = json['bluetooth_device_type'] as String
-    ..connectable = json['connectable'] as bool
-    ..txPowerLevel = json['tx_power_level'] as int
-    ..rssi = json['rssi'] as int;
+  return BluetoothDevice(
+    advertisementName: json['advertisement_name'] as String,
+    bluetoothDeviceId: json['bluetooth_device_id'] as String,
+    bluetoothDeviceName: json['bluetooth_device_name'] as String,
+    bluetoothDeviceType: json['bluetooth_device_type'] as String,
+    connectable: json['connectable'] as bool,
+    rssi: json['rssi'] as int,
+    txPowerLevel: json['tx_power_level'] as int?,
+  );
 }
 
 Map<String, dynamic> _$BluetoothDeviceToJson(BluetoothDevice instance) {
-  final val = <String, dynamic>{};
+  final val = <String, dynamic>{
+    'advertisement_name': instance.advertisementName,
+    'bluetooth_device_id': instance.bluetoothDeviceId,
+    'bluetooth_device_name': instance.bluetoothDeviceName,
+    'bluetooth_device_type': instance.bluetoothDeviceType,
+    'connectable': instance.connectable,
+  };
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -78,24 +83,21 @@ Map<String, dynamic> _$BluetoothDeviceToJson(BluetoothDevice instance) {
     }
   }
 
-  writeNotNull('advertisement_name', instance.advertisementName);
-  writeNotNull('bluetooth_device_id', instance.bluetoothDeviceId);
-  writeNotNull('bluetooth_device_name', instance.bluetoothDeviceName);
-  writeNotNull('bluetooth_device_type', instance.bluetoothDeviceType);
-  writeNotNull('connectable', instance.connectable);
   writeNotNull('tx_power_level', instance.txPowerLevel);
-  writeNotNull('rssi', instance.rssi);
+  val['rssi'] = instance.rssi;
   return val;
 }
 
 WifiDatum _$WifiDatumFromJson(Map<String, dynamic> json) {
-  return WifiDatum()
-    ..id = json['id'] as String
+  return WifiDatum(
+    ssid: json['ssid'] as String?,
+    bssid: json['bssid'] as String?,
+    ip: json['ip'] as String?,
+  )
+    ..id = json['id'] as String?
     ..timestamp = json['timestamp'] == null
         ? null
-        : DateTime.parse(json['timestamp'] as String)
-    ..ssid = json['ssid'] as String
-    ..bssid = json['bssid'] as String;
+        : DateTime.parse(json['timestamp'] as String);
 }
 
 Map<String, dynamic> _$WifiDatumToJson(WifiDatum instance) {
@@ -111,5 +113,6 @@ Map<String, dynamic> _$WifiDatumToJson(WifiDatum instance) {
   writeNotNull('timestamp', instance.timestamp?.toIso8601String());
   writeNotNull('ssid', instance.ssid);
   writeNotNull('bssid', instance.bssid);
+  writeNotNull('ip', instance.ip);
   return val;
 }