+++ packages/carp_connectivity_package/test/carp_connectivity_package_test.dart
@@ -3,27 +3,28 @@ import 'dart:io';
 import 'package:carp_connectivity_package/connectivity.dart';
 import 'package:test/test.dart';
 
+import 'package:carp_core/carp_core.dart';
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 
 String _encode(Object object) =>
     const JsonEncoder.withIndent(' ').convert(object);
 
 void main() {
-  CAMSStudyProtocol protocol;
+  late StudyProtocol protocol;
   Smartphone phone;
 
   setUp(() {
+    // make sure that the json functions are loaded
+    DomainJsonFactory();
+
     // register the context sampling package
     SamplingPackageRegistry().register(ConnectivitySamplingPackage());
 
     // Create a new study protocol.
-    protocol = CAMSStudyProtocol()
-      ..name = 'Context package test'
-      ..owner = ProtocolOwner(
-        id: 'AB',
-        name: 'Alex Boyon',
-        email: 'alex@uni.dk',
-      );
+    protocol = StudyProtocol(
+      ownerId: 'alex@uni.dk',
+      name: 'Context package test',
+    );
 
     // Define which devices are used for data collection.
     phone = Smartphone();
@@ -43,15 +44,15 @@ void main() {
   test('CAMSStudyProtocol -> JSON', () async {
     print(protocol);
     print(toJsonString(protocol));
-    expect(protocol.ownerId, 'AB');
+    expect(protocol.ownerId, 'alex@uni.dk');
   });
 
   test('StudyProtocol -> JSON -> StudyProtocol :: deep assert', () async {
     print('#1 : $protocol');
     final studyJson = toJsonString(protocol);
 
-    CAMSStudyProtocol protocolFromJson = CAMSStudyProtocol
-        .fromJson(json.decode(studyJson) as Map<String, dynamic>);
+    StudyProtocol protocolFromJson =
+        StudyProtocol.fromJson(json.decode(studyJson) as Map<String, dynamic>);
     expect(toJsonString(protocolFromJson), equals(studyJson));
     print('#2 : $protocolFromJson');
   });
@@ -60,10 +61,10 @@ void main() {
     // Read the study protocol from json file
     String plainJson = File('test/json/study_protocol.json').readAsStringSync();
 
-    CAMSStudyProtocol protocol = CAMSStudyProtocol
-        .fromJson(json.decode(plainJson) as Map<String, dynamic>);
+    StudyProtocol protocol =
+        StudyProtocol.fromJson(json.decode(plainJson) as Map<String, dynamic>);
 
-    expect(protocol.ownerId, 'AB');
+    expect(protocol.ownerId, 'alex@uni.dk');
     expect(protocol.masterDevices.first.roleName, Smartphone.DEFAULT_ROLENAME);
     print(toJsonString(protocol));
   });
@@ -74,13 +75,15 @@ void main() {
     print(_encode(dp));
 
     BluetoothDatum datum = BluetoothDatum()
-      ..scanResult.add(BluetoothDevice()
-        ..bluetoothDeviceId = "weg"
-        ..bluetoothDeviceName = "ksjbdf"
-        ..connectable = true
-        ..txPowerLevel = 314
-        ..rssi = 567
-        ..bluetoothDeviceType = "classic");
+      ..scanResult.add(BluetoothDevice(
+        advertisementName: 'myPhone',
+        bluetoothDeviceId: "weg",
+        bluetoothDeviceName: "ksjbdf",
+        connectable: true,
+        txPowerLevel: 314,
+        rssi: 567,
+        bluetoothDeviceType: "classic",
+      ));
 
     final DataPoint data = DataPoint.fromData(datum);