+++ packages/carp_context_package/example/lib/example.dart
@@ -1,4 +1,4 @@
-import 'package:carp_core/carp_protocols/carp_core_protocols.dart';
+import 'package:carp_core/carp_core.dart';
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 import 'package:carp_context_package/context.dart';